+++ protocol-units/bridge/chains/ethereum/src/types.rs
@@ -1,8 +1,26 @@
+use alloy::network::{Ethereum, EthereumWallet};
 use alloy::primitives::Address;
+use alloy::providers::fillers::{
+	ChainIdFiller, FillProvider, GasFiller, JoinFill, NonceFiller, WalletFiller,
+};
+use alloy::providers::RootProvider;
 use alloy::rlp::{RlpDecodable, RlpEncodable};
+use alloy::transports::BoxTransport;
 use serde::{Deserialize, Serialize};
 
 pub type EthHash = [u8; 32];
+pub(crate) type AlloyProvider = FillProvider<
+	JoinFill<
+		JoinFill<
+			JoinFill<JoinFill<alloy::providers::Identity, GasFiller>, NonceFiller>,
+			ChainIdFiller,
+		>,
+		WalletFiller<EthereumWallet>,
+	>,
+	RootProvider<BoxTransport>,
+	BoxTransport,
+	Ethereum,
+>;
 
 #[derive(Debug, PartialEq, Eq, Hash, Clone, RlpEncodable, RlpDecodable, Serialize, Deserialize)]
 pub struct EthAddress(pub Address);
