+++ protocol-units/bridge/chains/ethereum/src/utils.rs
@@ -1,31 +1,13 @@
 use std::str::FromStr;
 
 use crate::types::EthAddress;
-use alloy::contract::{CallBuilder, CallDecoder};
-use alloy::network::{Ethereum, EthereumWallet};
 use alloy::primitives::U256;
-use alloy::providers::{
-	fillers::{ChainIdFiller, FillProvider, GasFiller, JoinFill, NonceFiller, WalletFiller},
-	Provider, RootProvider,
-};
 use alloy::rlp::{Encodable, RlpEncodable};
-use alloy::rpc::types::TransactionReceipt;
-use alloy::transports::{BoxTransport, Transport};
 use keccak_hash::keccak;
+use mcr_settlement_client::send_eth_transaction::{
+	InsufficentFunds, SendTransactionErrorRule, UnderPriced, VerifyRule,
+};
 use thiserror::Error;
-
-pub(crate) type AlloyProvider = FillProvider<
-	JoinFill<
-		JoinFill<
-			JoinFill<JoinFill<alloy::providers::Identity, GasFiller>, NonceFiller>,
-			ChainIdFiller,
-		>,
-		WalletFiller<EthereumWallet>,
-	>,
-	RootProvider<BoxTransport>,
-	BoxTransport,
-	Ethereum,
->;
 
 #[derive(Debug, Error)]
 pub enum EthUtilError {
@@ -54,21 +36,6 @@ impl FromStr for EthAddress {
 	}
 }
 
-pub async fn send_transaction<
-	P: Provider<T, Ethereum> + Clone,
-	T: Transport + Clone,
-	D: CallDecoder + Clone,
->(
-	contract_call: CallBuilder<T, P, D, Ethereum>,
-) -> Result<TransactionReceipt, EthUtilError> {
-	let pending_transaction = contract_call.send().await?;
-
-	pending_transaction
-		.get_receipt()
-		.await
-		.map_err(|e| EthUtilError::GetReceiptError(e.to_string()))
-}
-
 pub fn calculate_storage_slot(key: [u8; 32], mapping_slot: U256) -> U256 {
 	#[derive(RlpEncodable)]
 	struct SlotKey<'a> {
@@ -83,4 +50,10 @@ pub fn calculate_storage_slot(key: [u8; 32], mapping_slot: U256) -> U256 {
 
 	let hash = keccak(buffer);
 	U256::from_be_slice(&hash.0)
+}
+
+pub(crate) fn send_tx_rules() -> Vec<Box<dyn VerifyRule>> {
+	let rule1: Box<dyn VerifyRule> = Box::new(SendTransactionErrorRule::<UnderPriced>::new());
+	let rule2: Box<dyn VerifyRule> = Box::new(SendTransactionErrorRule::<InsufficentFunds>::new());
+	vec![rule1, rule2]
 }
